rework import process
